Code review comments + minor renames.
