+++ Vostok.Core/Flow/Serializers/BytesArraySerializer.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Vostok.Flow.DistributedContextSerializer
+namespace Vostok.Flow.Serializers
 {
     public class BytesArraySerializer : BaseTypedSerializer<byte[]>
     {
