+++ Vostok.Core/Flow/Serializers/GuidSerializer.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Vostok.Flow.DistributedContextSerializer
+namespace Vostok.Flow.Serializers
 {
     internal class GuidSerializer : BaseTypedSerializer<Guid>
     {
